FAB-14502 Clarify deliver 'Errored' message
The deliver service is currently quite ambiguous about the error
condition when the channel returned by Errored() closes.  This is likely
because this message could be returned by a peer or orderer.  However,
in reality, the only time this condition ever occurs is because of a
consenter error.  In the interest of serviceability, this CR makes the
error message more specific, making a note that we should handle
generalizing it if the peer ever implements this function.

Change-Id: I5b2db28c89876745d600a5191a7d0cfa95a97c56
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>