[FAB-14578] Sane defaults for block cutting
A max of 10 transactions per block is too low because
it creates tiny blocks. We should increase it to a
saner value which is more optimized for throughput
instead of letting the users always do it themselves.

Correspondingly, 256 blocks circulating the consensus
is too high and should be tuned down appropriately.

Change-Id: Ib2f2c0f9c1875102e14d73f054f67103036b3d8d
Signed-off-by: yacovm <yacovm@il.ibm.com>