[FAB-14573] install chaincode returns a package ID
The chaincode install steps returns an identifier of the package which is
constructed as the concatenation of the chaincode package label, a separator
(:) and the package hash. This can later be used in the approve step to
signal to peers of an org which chaincode package to run upon invocation.

Change-Id: If2e5d07e2fbf8210c1c17d2896eb1f6fc56d932b
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>