rollback: disable reconciler during block reprocessing
After a rollback, we need to quickly get the peer to
a pre-rollbacked block height by reprocessing blocks.

During this period, we can disable the
reconciler to avoid interference with the block
reprocessing (would help in reducing the time taken
to reprocess blocks). Further, the missing pvtData
info in the pvtdataStore could belong to a block
which is yet to be processed. In this case, we cannot
run reconciler and try to update the stateDB. Otherwise,
it would end up in a inconsistent stateDB.

The missing pvtData reconcilation is disabled by
returning nil missingDataInfo.

FAB-14315 #done

Change-Id: I205fc84cea09c381c3aa9ef6677e3ab7b7e9659c
Signed-off-by: senthil <cendhu@gmail.com>
(cherry picked from commit a16f93138c29b7bb4b4b90f4da0af86eb7078763)