FAB-14551 Properly decorate org proto
After the split of the proto annotations and the protos, some consumers
of protolator were broken.  configtxgen is one of those consumers,
particularly, its printOrg function assumed that the vanilla proto had
annotations attached to it which allowed the opaque types to be resolved
correctly.  It did not, so it output bytes, but configtxlator did which
would not allow unmarshaling of those bytes.  This CR simply adds the
appropriate annotation.

Change-Id: Iaa0692b709a1fbdb27a6e0e049ef2490e1d5210b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>