[FAB-14621] CCID/PackageID refactoring
This change set creates a new type (PackageID, under persistence) and makes
use of it instead of CCID wherever appropriate. The idea is that runtime and
persistence each have their own identifier for the chaincode/package
artefact.  Currently they are each backed by the same string, but this may
change in the future. Also, both types implement fmt.Stringer so that it's
not necessary to cast back to the underlying string to obtain a string
representation.

Change-Id: Ide2918f354e47f9ee49e92de4cd245d97d185bb8
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>