FAB-14656 Respect snapshot interval when node restarts
When a node is restarted, it loads WAL data since last snapshot
into memory, however this chunk of data is not currently taken
into account as part of accumulated data, which is compared to
snapshot interval.

This CR fixes this. Also it changes some log level to improve
serviceability.

FAB-14656 #done

Change-Id: If152071e64fd8268d20362c593d24af4ab2be355
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>