[FAB-14587] latest ch. config to evaluate policies
Problem: the defect shows the inability to evaluate a channel-scoped policy
for an organisation which is only added after genesis. Policy evaluation is
implemented by a policy manager. The root cause of the problem is that a
reference to the policy manager is obtained once at start-up time and never
refreshed: as a consequence, the channel configuration may change but the
reference to the policy manager will not. In the defect, we see this happen:
the policy manager is only aware of two organisations, and when a third one
is added in a later config update, the old policy manager reference cannot
evaluate any policy for this new org.

Solution: the same approach used for plugins is adopted: a dynamic manager is
used. The dynamic manager retrieves a fresh reference to the channel first and
then carries on to the evaluation.

Change-Id: Icaa30dbe2b5acc22e7a55d22bbb95a6d4afddc13
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>