[FAB-14687] Only add to puller if needed
The gossip channel's method AddToMsgStore adds a block message
to both the block store and to the block puller.

However, it can be that the message isn't added to the block store
due to it being too old.

In such a case we must not add it to the block puller, because it is evicted
by the invalidation trigger of the block store which is programmed to
remove it from the block puller.

Change-Id: I64e7f4e64a337df5593a58a5d52a821b376b2bfb
Signed-off-by: yacovm <yacovm@il.ibm.com>