[FAB-14688] Check inner, not outer block msg
The gossip code that handles pull block data updates
has an optimization that first performs a simulation
of adding an inner block message to the message store,
and only if the simulation is successful  -
it then verifies the message.

The problem is that the simulation is done against the outer
message, and thus the optimization does not occur.

This change set fixes this.

Change-Id: I71c94e9b1cad2bfa6e9cfce700d8bd3209e9b5e7
Signed-off-by: yacovm <yacovm@il.ibm.com>