[FAB-14696] Proper lastConfigMetadata in inactiveChain
This change set properly sets the last config metadata for inactive chains.
Inactive chains are chains that the orderer isn't part of, but still has the
genesis block in its ledger, to avoid state forks.

The replication code artificially creates a genesis block for the channel from
the channel creation transaction.

Change-Id: I57a4fbb609ae447d25041e64aaf71e151e83f68f
Signed-off-by: yacovm <yacovm@il.ibm.com>