FAB-14700  Kafka2Raft validate broadcast
Add methods to validate a config update in broadcast phase.
Validate a proposed config update versus the current config.

Checks the validity of the state transitions of a possible
migration config update tx by comparing the current
ConsensusType config with the next (proposed)
ConsensusType fields. It is called during the broadcast phase
and never changes the status of the underlying chain.

Change-Id: Ib9b7d75f703d7f7f5eb0c3d514e2bfd4073f34f9
Signed-off-by: Yoav Tock <tock@il.ibm.com>