FAB-14735 Ignore unchanged consenters in update
The config update detection code in Raft is a re-implementation of the
fabric config checking mechanisms.  This should probably be changed in
the long term, but one of its currently unhandled cases, is when a
config update references the consenters in the write set, but does not
modify them.  This can occur especially when adding a new organization
to the orderer group.  This CR adds the additional version checks
necessary to detect and ignore these sorts of updates.

Change-Id: Ib35a97e1cdbd557705f11da183c8547f9f85539a
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
(cherry picked from commit 59c3387d0d6fa32f4af895eb0483e1d385bd401f)
Signed-off-by: Artem Barger <bartem@il.ibm.com>