Fix error on peer restart
Upon peer restart after defining a chaincode, lifecyle
cache popluation leads to an error. Because, the this
cache is populated via a state listener initialize which
gets invoked during ledger opening. The implementaiton
of this function in turns need access to channel config
that gets loaded after ledger opening. As as fix, in this
CR, the validation of collection names is avoided to break
this cycle.

Though this is not a problem but proper fix would be desired
where initialization of other components does not depend on
internal steps of ledger opening. Rather, the initilization
should explicitly take place in the desired order outside ledger.
However, this would need a bigger change in the interaction of
state listeners and ledger and requires some more thinking - so
this cr for an intrim fix for the problem.

FAB-14742 #done

Change-Id: I6df2a8bce19243899afef44bbaa2b51c3418dfbe
Signed-off-by: manish <manish.sethi@gmail.com>