Explicitly set ext key usage for CC CA
The CA cert generated by tlsgen which
is used by the chaincode server
currently sets ExtKeyUsage to
ExtKeyUsageAny.  This is actually not
allowed by the spec and is now enforced by
openssl 1.1.0 and later.  The main issue
for Fabric is that without this change
Node CC cannot move to Node 10.

Change-Id: Ifd4a5733ca4febb44c5c36943b78a4f82e657906
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>