FAB-14735 (fixed IT) ignore not changed consenters
This fix adds IT *properly initialized and finalized*  to check
addition of new orderer organization and fix the code path where
etcd/chain.go checks for consenters set validity. Moreover this commmit
adds code to check the leader is elected so to attest system channel is
actually ready to pull configuration block from.

Change-Id: If3cb9c1313229b67d2d63669aa663dea1080fdd3
Signed-off-by: Artem Barger <bartem@il.ibm.com>