[FAB-14770] move async behavior to caller
The logic of handleInit and handleTransaction were contained completely
with in a closure that was launched asynchronously as a go routine.

Instead of hiding it in the impelmentation, just make the (single)
callers start the implementation in a go routine. This cleans up the
implementation and will make it easier to test it explicitly in a
synchronous manner when we refactor tests later.

Change-Id: I50e5c94325e62121f2bd2a5293bccbd39493d7e9
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>