[FAB-14770] sendReceive is synchronous
The implementation would send a message in a go routine and wait for the
send to complete concurrent with the wait for a response. That pattern
is unnecessary though, as the response will never come if the message
was not sent successfully.

Simplify by moving to serialSend and handling the error before moving on
to wait for the response.

Change-Id: I0326c5583ac8c693bcfb342600211dbfdb3b3aa9
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>