[FAB-14914] translator for SBE policies
State-based endorsement policies are serialised as SignaturePolicyEnvelope
policies. However the v20 validator interprets all policies as instances of
ApplicationPolicy. While in future release we plan to package chaincodes that
produce ApplicationPolicy directly, to use a new meta key to disambiguate
policies and to transparently replace all SignaturePolicyEnvelope policies with
their ApplicationPolicy counterpart, for now we simply add a level of
translation from SignaturePolicyEnvelope to ApplicationPolicy. This CR also
adds an integration test that exercises state-based endorsement with the new
lifecycle.

Change-Id: I89dd253ecdbb6f08935fc1795edadabfcccdf6dd
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>