[FAB-14818] Flake in TestStreamAbort
This change set fixes a flake in TestStreamAbort by ensuring
the server-side handler doesn't return before the stream
is canceled by the reconfiguration.

This is needed, because gRPC selects the error to return
to the Recv() method by doing a select { } between
the actual bytes read from the stream, and the context.

In the test, we test that the context is canceled due to
the reconfiguration, but we also concurrently let the server
side handling of the request finish - it might be that the
bytes read (and the corresponding EOF signal) happens before
the context cancelling, and therefore the expected error
is not received.

By delaying the server side handling of the request, we ensure
that we either abort the stream, or the test never ends.

I ran it (with data race detector) 5,000 times and it succeeded.

Change-Id: Ie1ecf360dde79c3afa7bf919a84a20b3b600c865
Signed-off-by: yacovm <yacovm@il.ibm.com>