[FAB-14931] refresh cc cache on packageID update
This change set adds the hash of the key that stores the package ID to the
set of interesting hashes that are tracked by the update mechanism of the
chaincode cache. This way, the admin can later update the definition by
specifying a different (or empty) package ID as a means to temporarily disable
endorsements for that chaincode by peers of that org. Integration tests have
been extended to verify the scenario.

Change-Id: Ifc64897a26395cb46aaccd5c7194cada3a070639
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>