rollback: skip uncommittedPvtData of oldBlocks
As we are storing pvtData of both valid and invalid tx
in the pvtdataStore, we do store missing info for both
valid and invalid tx. When the reconciler fetches the
missing pvtData, it passes to the ledger coordinator.
As the pvtData contains data of both valid and invalid tx,
we need to pass both to pvtdataStore and only valid tx's
pvtData to the txmgr.

Similarly, during failure and recovery, we fetch the pvtData
of last updated old blocks and filter uncommitted tx pvtData
while syncing the stateDB

FAB-14861 #done

Change-Id: I858939beba2c702df7f197aa6a6c555f4655a83b
Signed-off-by: senthil <cendhu@gmail.com>
(cherry picked from commit df3616a02e4e67886dc78aa42520b40524390cc9)