[FAB-14942] fix membership tracker test
When CI is slow it might take few miliseconds to consume a message from
the stop channel, therefore in rare cases membership tracker might do an
extra check before it's stopped. This commit changes unit test to remove
assertion of amount of time it's being called, raplacing with the
assertion that it won't be executed once stopped.

Change-Id: I0effc5a7e05df367f36840a55bbf696d81fb5cea
Signed-off-by: Artem Barger <bartem@il.ibm.com>