[FAB-15181] Introduce consensus.MetadataValidator
This CR introduces a consensus.MetadataValidator interface that must be
implemented by a consensus type implementation to validate updates to
its metadata. We also provide a no-op implementation that can be used by
consensus types that do not need any such validation (like solo and
kafka). The MetadataValidator is independent of the Consenter or the
Chain implementation since given the old metadata we should be able to
validate a new metadata without need for any further state information.

Change-Id: Ifcea866a096d97c1fd858f2689362b91fc028d03
Signed-off-by: Adarsh Saraf <adarshsaraf123@gmail.com>