FAB-14401 Pass GO_TAGS to Docker targets
Enabling certain features (e.g. pkcs11) in the
peer and orderer binaries requires Go
build tags.  This change passes GO_TAGS
to the peer and orderer Docker targets.

Given the goal was to allow people to build
images with pkcs11 enabled, also needed to add
the libtool package to the golang intermediate
images used in the Dockerfiles.

FAB-14401 #done

Change-Id: Iba58f35c4efc13b1ebdc8e2b6792d07eae74211d
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>