[FAB-15143] Fix Test flake in TestDifferentMessages
The test tries to evict entries from the signer cache by inserting new
ones (different than the existing ones) in parallel, however it can be
that all insertions reach the conclusion that there is no need to shrink
the cache (because of concurrency).

Therefore we:
1) Make the insertion serial instead of parallel
2) Increase the amount of items we opreate on by 10,
in order for the probability of no one of them to be picked to be negligible.

I ran the test 50,000 times and it doesn't fail now.

Change-Id: I2de7b9f17cc6b27ffda2af11d1f5b3e8b33f3a67
Signed-off-by: yacovm <yacovm@il.ibm.com>