[FAB-15238] legacy cc defs must not require init
Chaincodes defined via LSCC must not return true in their implementation
of the RequiresInit function of the ChaincodeDefinition interface. The
Init behaviour for the legacy lifecycle is anyway handled through
ExecuteLegacyInit irrespective of this function. On the contrary, if
RequiresInit returns true, as soon as the channel is upgraded to 2.0,
invocations of chaincodes defined via LSCC will fail because the new
init-enforcing logic will not find the expected key that governs the
presence of init and will fail the endorsement.

Change-Id: Ibe872a75c4f9e32d8b0809ea5e3b4d1901b729cd
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>