[FAB-15261] Fix cc2cc discovery query bug
Given the following 2 endorsement policies (in any order),
the chaincode to chaincode query return an empty set:

policy 1: AND(Org1.member, Org2.member)
policy 2: OR(Org1.member, Org2.member)

This is due to the Reduce() function unable to handle
principal sets that are identical, and thus it reduces it
to an empty set.

Fixed by adding a check to the Reduce() function that if
2 principal sets cover each other, it leaves one of them
in the result (by their relative order).

This change also requires to sometimes re-evaluate the reduction
because it may leave principal sets that can be further reduced,
so I added some logic that ensures the reduction is made
until no more improvement can be made.

Change-Id: I328ed338368f92efa81ca4459987997c4d21943b
Signed-off-by: yacovm <yacovm@il.ibm.com>