[FAB-15267] Refactor goGen in discovery/api
Go Generate in discovery/api that generates the mocks
for GossipSupport seems to be generated in the wrong
file with output directory path incorrectly. (output
path was not being used anywhere).

The new GossipSupport mock generation is moved into
discovery/support/support.go which generate the mocks
for GossipSupport via local interface and output to
discovery/support/mocks, which is being used in
several tests.

Change-Id: Ic85c36cf2bd24512b5537581465da6e1682f6ab5
Signed-off-by: Chongxin Luo <Chongxin.Luo@ibm.com>