FAB-15173 Consensus migration: maintenance filter
Entry into maintenance mode should not be accompanied with
additional config changes, because the whole point of maintenance
mode is to first create a stable and safe backup point, with a
good configuration. In particular, it should not be allowed change
the consensus type together with entry or exit from maintenance mode.
Changing the consensus type should only be done in maintenance mode.

In addition the type of consensus type transitions should be
checked - currently we only support kafka to raft.

- Transition in or out of maintenance mode (i.e. a change in
  ConsensusType.State) cannot be accompanied with a chenge to any
  othe group or value except /Channel/Orderer/ConsensusType.

- Entry: Prevent transition into maintenance mode together with
  change in ConsensusType.Type
- Prevent changes to ConsensusType.Type unless in maintenance mode
- Exit: Prevent transition out-of maintenance mode together with
  change in ConsensusType.Type

This is achieved by adding a new filter, the maintenance-filter,
which is only evaluated on ProcessesConfigUpdate(). This is unlike
all other filters, which are evaluated on both normal and config
messages. For that end it is added to a separate Rule field in the
standardchannel.go. This Rule can later be extended to
a <configFilters []Rule> array (slice) if we find we need more
filters that apply to config only.

When capability OrdererV1_4_2 (or later) is off, entry into
maintenance mode or changing consensus-type should be prevented.

Change-Id: Ib9bd823042b71d972739de1c1a78f6343dbd794f
Signed-off-by: Yoav Tock <tock@il.ibm.com>