[FAB-15457] Fix a typo "networkID" in core.yaml
The names of the docker images and containers of Chaincodes do
not have proper prefixes.
The root cause is a typo in fabric/sampleconfig/core.yaml,
which is used for peer docker images. "networkId" is the correct
parameter. However, core.yaml has a wrong parameter "networkID".
This CR fixes the letter "D" to a lower case.

FAB-15457 #done

Change-Id: I880c22642c7f49f91d207facd5bb127b24a45306
Signed-off-by: Yuki Kondo <yuki.kondo@hal.hitachi.com>