[FAB-15450] validate historydb key if it contains nils
There was a logic to skip keys with nil bytes from the historydb results.
However, because some valid blockNumTranNumBytes may also contain
nil bytes, these history keys are filtered out.

This CR updates the logic to handle decoding error and process all keys
returned by range query no matter if blockNumTranNumBytes contains
a nil byte. If there is a decoding error, block:tran cannot be found,
or key is not present in write set, it means that the key is a false key
and will be skipped.

Change-Id: I296479af55b7c750ef8f797faa807ee0f5489b0a
Signed-off-by: Wenjian Qiao <wenjianq@gmail.com>