FAB-15376 fix sizefilter: take dynamic resources
The size filter was given the orderer config at creation, rather
than a link to the channel resources.

The sizefilter needs to get a new orderer config on every
invocation of Apply, in order to keep up with dynamic
configuration changes. This is done using the link to the channel
resources (as in other filters, e.g. sigfilter.go).

Change-Id: I8ca6c1e52b44e1c43e88f929c9491bd90b7eb53c
Signed-off-by: Yoav Tock <tock@il.ibm.com>