blockstore computes txID if it is empty
Currently, a config transaction (both genesis and config
update) might have an empty txID. As the genesis block and
the subsequent config update block do not have the txID,
a collision would occur at the blockstore index. As we
check for duplicate index before creating index for each
tx, we skip it.

This CR ensures that the correct txID is computed if
it is empty so that any kind of collision would be
minimized. Note that it can only be minimized as
validator does not check for duplicate txID in
config transaction. Hence, there is a possibility of
collisions in index store.

FAB-15651 #done

Change-Id: I751b591e8e55c7a411a0cfa68451a70dcef96cb8
Signed-off-by: senthil <cendhu@gmail.com>