FAB-15654 Undo v2.0 capability backport
During the raft migration backport, the v2_0 capability was accidentally
backported along with it.  This does create the potential for a fork in
the v1.4.1 release, but an incredibly unlikely one.  The orderer admins
would have to explicitly enable the v2.0 capability, then use a v2.0
version of configtxgen, with the `-baseProfile` option, while some v2.0
orderers were also active on the network.  Regardless, as this is a bug,
it should be fixed in v1.4.2.

Change-Id: Ib3b0b832c5a7c21bd1ab11b1694314c4b6d643b5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>