[FAB-14324] remove peer.channelPolicyManagerGetter
The function adapter and constructor would always return a non-nil policy
manager and true or a nil policy manager and false based on whether or
not a policy manager could be obtained for a channel.

The calling code would generally ignore the flag and only check for the
nil or non-nil manager for decision making.

To remove the wrappers and simplify, the
policies.ChannelPolicyManagerGetter interface was changed to remove the
bool return value and a simple function adapter was introduced to cast
the policy manager returned by the peer into a policies.Manager.

Change-Id: Iacd2cb5c366dbc8f67da7e2fd9515ae52e36c64f
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>