[FAB-15732] Remove shim and protos filter in platform
Currently during the chaincode build process the shim
and peer protos are not added to the dependency tree as
they are part of the ccenv image. The ccenv image will be
changed in a future changeset to remove the inclusion of
fabric components from it, this change is preparation for that.

Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>
Change-Id: I5819a126c601f99ba783735bbf77097ed23ea24f