code hygine: rename addBlockCommitHashIfApplicable()
IfApplicable in addBlockCommitHashIfApplicable() makes the
function bit long and might be unnecessary too.  Hence,
this CR remove ifApplicable and make it addBlockCommitHash().
As ifApplicable is removed, the condition is moved out the
function so that addBlockCommitHash() always adds the commits
hash to the block.

FAB-15788 #done

Change-Id: I53fd7f12c52d17ef3565d9f60d68fe5d7b4d7d98
Signed-off-by: senthil <cendhu@gmail.com>
(cherry picked from commit c253e10374f6454e3311998927f4275410466eef)