tests for commit hashes with rollback and reset
The COMMIT_HASH is being added to each block both
in v1.1 and >= v1.4 As this hash has to be the same
on all peers, it is necessary to start adding this
hash from the same block on all peers. Hence, it is
necessary to reset the existing channels to the
genesis block inorder to add COMMIT_HASH. Otherwise,
the peer would not add COMMIT_HASH to blocks.

There are two new sets of tests:
(1) ensure that the COMMIT_HASH is being added to
the block once a peer has been reset to the genesis
block. Otherwise, COMMIT_HASH would not be added.
(2) assuming that COMMIT_HASH is already present,
doing a rollback/reset and reprocessing the block
should not result in different COMMIT_HASH (as we
assume no changes in the validation results). This
test also helps to ensure that any changes to the
code that computes the COMMIT_HASH do not break
the backward compatibility.

FAB-17832 #done

Signed-off-by: manish <manish.sethi@gmail.com>
Signed-off-by: senthil <cendhu@gmail.com>