[FAB-15933] Optional NodeOU Classification
Currently, once NodeOU classification is enabled, it is enforced
that client, admin, peer, and orderer classification are
properly specified. Though, not all organisations are required
to have support for all classifications.
Therefore, even when NodeOU classification is enabled,
specific classifications can be omitted.
The expected behaviour should be the following:
1. if a classification is not provided then the setup
should not fail.
2. if a classification is not provided testing an
identity against that classification will fail.
3. if but no valid classification is provided, then
NodeOU classification is disabled.
4. if no valid admin classification is provided and
the no msp admins are declared, the setup will fail.

Change-Id: Ie2a69a15e7ae95868dec5135d8e76ee5625bbbc2
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: yacovm <yacovm@il.ibm.com>