filelock to allow one of peer start, reset, rollback
As of now, the admin has to be careful to issue only one
of the instance of one of the commands (start, reset, and
rollback) at a time in order to prevent a chance of data
corruption. However, these three should be made mutually
exclusive in the code in order to prevent this mistake.

A simple filelock on a file in the ledger's dir can be
used for this purpose

FAB-15923 #done

Change-Id: Iaa5fc31a01b960c40d234d9b9b49435a28eb656b
Signed-off-by: senthil <cendhu@gmail.com>
(cherry picked from commit b137f5be2e8610c72eab05c881131572445eebd6)