FAB-15935 Make ConnectionTimeout configurable
The orderer currently has no way to override the ConnectionTimeout gRPC
parameter.  This CR makes the variable configurable via the orderer.yaml
or environment.  Because this parameter usually does not need to be
overridden, it is omitted from the configuration file.

The peer also has no way, so the CR also adds one for the peer.

Change-Id: Ifb93d151a83daabcf0086e1019fe4c3607f99161
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>