FAB-14491 Remove pointless metadata routing
All of the existing chaincode types use the exact same metadataprovider,
so it's not entirely clear why we make this a pluggable point.  Getting
rid of the pluggability.

Note, this eliminates a test, and decreases testing code coverage.  Note
however that it eliminates the exact same untested code in other places.
This code will be further eliminated in a future CR.

Change-Id: I84ec7fdf4c131ce576a8b607d8eef893758a1683
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>